+++ src/js/modules/activities/index.ts
@@ -1,6 +1,7 @@
 import * as actions from './actions';
 import reducer from './reducer';
+import createSagas from './sagas';
 import * as selectors from './selectors';
 
-export default { actions, reducer, selectors };
+export default { actions, reducer, selectors, createSagas };
 export { Activity, ActivityState, ActivityType, LoadActivitiesAction, LoadActivitiesForProjectAction } from './types';
