+++ src/js/modules/branches/index.ts
@@ -1,8 +1,9 @@
 import * as actions from './actions';
 import reducer from './reducer';
+import createSagas from './sagas';
 import * as selectors from './selectors';
 
-export default { actions, reducer, selectors };
+export default { actions, reducer, selectors, createSagas };
 export {
   Branch,
   BranchState,
