+++ src/js/modules/commits/index.ts
@@ -1,6 +1,7 @@
 import * as actions from './actions';
 import reducer from './reducer';
+import createSagas from './sagas';
 import * as selectors from './selectors';
 
-export default { actions, reducer, selectors };
+export default { actions, reducer, selectors, createSagas };
 export { Commit, CommitState, LoadCommitsForBranchAction } from './types';
