+++ src/js/modules/forms/index.ts
@@ -1,8 +1,11 @@
-import { Action, Dispatch } from 'redux';
+import { Dispatch } from 'redux';
+
+export { default as createSagas } from './sagas';
 
 // Based on https://github.com/yelouafi/redux-saga/issues/161#issuecomment-229350795
 export const FORM_SUBMIT = 'FORMS/FORM_SUBMIT';
-export interface FormSubmitAction extends Action {
+export interface FormSubmitAction {
+  type: 'FORMS/FORM_SUBMIT';
   payload: {
     submitAction: string;
     successAction: string;
