+++ src/js/modules/user/index.ts
@@ -1,8 +1,9 @@
 import * as actions from './actions';
 import reducer from './reducer';
+import createSagas from './sagas';
 import * as selectors from './selectors';
 
-export default { reducer, selectors, actions };
+export default { reducer, selectors, actions, createSagas };
 export {
   Team,
   UserState,
