+++ src/js/sagas/utils.ts
@@ -126,6 +126,8 @@ export const createCollectionFetcher = <ApiParams>(
   };
 };
 
+debugger;
+
 const storingMetadata: {
   type: ApiEntityTypeString,
   storeActionCreator: (entities: EntityType[]) => StoreEntityAction,
